api/pyproject.toml
[tool.poetry]
name = "yogi-api"
version = "0.1.0"
description = "Assessment to show basic API skills."
authors = ["John Oatey"]
readme = "README.md"
packages = [{include = "yogi_api"}]

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = {extras = ["debug-server", "fastapi"], version = "^0.142.3"}
sqlalchemy = "^1.4.44"
asyncpg = "^0.27.0"

[tool.poetry.scripts]
my-script = "my_package.log_revision:start"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
